[Bronze II] Title: Hashing, Time: 100 ms, Memory: 14184 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -7,11 +7,11 @@ public class Main {
 		// 문자열의 길이 L
 		int L = Integer.parseInt(br.readLine());
 		// 해쉬 값
-		int hash = 0;
+		long hash = 0;
 
 		for(int i=0; i<L; i++) {
 			char tmp = (char) br.read();
-			hash += (int) (((tmp-'a'+1) * Math.pow(31, i)) % 1234567891);
+			hash += (long) (((tmp-'a'+1) * Math.pow(31, i)) % 1234567891);
 		}
 
 		System.out.println(hash);
